+++ src/Application/Services/CalendarService.cs
@@ -3,6 +3,7 @@ using Application.Errors.Services;
 using Application.Features.Requests;
 using Application.Features.Responses;
 using Application.Interfaces.Application;
+using Application.Interfaces.Application.Common;
 using Application.Interfaces.Infrastructure.Logging;
 using Application.Interfaces.Infrastructure.Services;
