+++ src/Application/Services/Common/DateTimeService.cs
@@ -1,6 +1,6 @@
-﻿using Application.Interfaces.Infrastructure.Services;
+﻿using Application.Interfaces.Application.Common;
 
-namespace Infrastructure.Services;
+namespace Application.Services.Common;
 
 /// <summary>
 /// The date time service class.
@@ -11,6 +11,7 @@ namespace Infrastructure.Services;
 internal sealed class DateTimeService : IDateTimeService
 {
 	public DateTime Today => DateTime.Today;
+	public TimeSpan TimeOfDay => DateTime.Now.TimeOfDay;
 	public DateTime Now => DateTime.Now;
 	public DateTime UtcNow => DateTime.UtcNow;
 }
