+++ src/Infrastructure/Services/AuthenticationService.cs
@@ -6,6 +6,7 @@ using System.Text;
 using Application.Contracts.Requests.Identity;
 using Application.Contracts.Responses.Identity;
 using Application.Errors.Services;
+using Application.Interfaces.Application.Common;
 using Application.Interfaces.Infrastructure.Logging;
 using Application.Interfaces.Infrastructure.Services;
