+++ src/Presentation/Controllers/CardController.cs
@@ -2,7 +2,7 @@
 
 using Application.Contracts.Requests.Finance;
 using Application.Contracts.Responses.Finance;
-using Application.Interfaces.Application;
+using Application.Interfaces.Application.Finance;
 using Application.Interfaces.Presentation.Services;
 
 using Asp.Versioning;
