+++ tests/ApplicationTests/Services/TransactionServiceTests.cs
@@ -3,7 +3,7 @@ using Application.Contracts.Responses.Finance;
 using Application.Errors.Services;
 using Application.Features.Requests;
 using Application.Features.Responses;
-using Application.Interfaces.Application;
+using Application.Interfaces.Application.Finance;
 
 using BaseTests.Helpers;
